// src/modules/garmentsProduction/reports/sewing/onlne-display-board/online-display-board-type.ts
export interface IOnlineDisplayBoard {
  LINENAME: string;
  LINEINCHARGE: string;
  BUYERNAME: number;
  STYLENO: string;
  SMVSEWING: number;
  SEWINGOUTPUT: number;
  EARNINGMIN: number;
  AVAILABLEMIN: number;
  EFFICIENCY: number;
  TARGET: number;
  TOTAL_SEWINGOUTPUT: number;
  TOTAL_EARNINGMIN: number;
  TOTAL_AVALIABLEMIN: number;
  TOTAL_EFFICIENCY: number;
  DAY_TARGET: number;
  DHU: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  COMPANY_REMARKS: string;
}